+++ mastracode/src/agents/subagents/index.ts
@@ -1,16 +1,12 @@
 /**
  * Subagent registry — maps subagent IDs to their definitions.
  */
-
-export type { SubagentDefinition } from "./types.js"
-export { exploreSubagent } from "./explore.js"
-export { planSubagent } from "./plan.js"
-export { executeSubagent } from "./execute.js"
-
-import type { SubagentDefinition } from "./types.js"
+import { createViewTool, createExecuteCommandTool, createGrepTool, createGlobTool, createWriteFileTool, createSubagentTool, stringReplaceLspTool, todoWriteTool, todoCheckTool } from "../../tools/index.js"
+import { resolveModel } from "../model.js"
+import { executeSubagent } from "./execute.js"
 import { exploreSubagent } from "./explore.js"
 import { planSubagent } from "./plan.js"
-import { executeSubagent } from "./execute.js"
+import type { SubagentDefinition } from "./types.js"
 
 /** All registered subagent definitions, keyed by ID. */
 const subagentRegistry: Record<string, SubagentDefinition> = {
@@ -34,4 +30,49 @@ export function getSubagentDefinition(
  */
 export function getSubagentIds(): string[] {
 	return Object.keys(subagentRegistry)
+}
+
+
+export function getSubagentTools(projectPath: string) {
+	// Create tools with project root
+	const viewTool = createViewTool(projectPath)
+	const executeCommandTool = createExecuteCommandTool(projectPath)
+	const grepTool = createGrepTool(projectPath)
+	const globTool = createGlobTool(projectPath)
+	const writeFileTool = createWriteFileTool(projectPath)
+
+	// The subagent tool needs tools and resolveModel to spawn subagents.
+	// We pass all tools that subagents might need based on their type.
+	const subagentTool = createSubagentTool({
+		tools: {
+			// Read-only tools (for explore, plan)
+			view: viewTool,
+			search_content: grepTool,
+			find_files: globTool,
+			// Write tools (for execute)
+			string_replace_lsp: stringReplaceLspTool,
+			write_file: writeFileTool,
+			execute_command: executeCommandTool,
+			// Task tracking (for execute)
+			todo_write: todoWriteTool,
+			todo_check: todoCheckTool,
+		},
+		resolveModel,
+	})
+
+	// Read-only subagent tool for plan mode — no execute type allowed
+	const subagentToolReadOnly = createSubagentTool({
+		tools: {
+			view: viewTool,
+			search_content: grepTool,
+			find_files: globTool,
+		},
+		resolveModel,
+		allowedAgentTypes: ["explore", "plan"],
+	})
+
+	return {
+		tool: subagentTool,
+		toolReadOnly: subagentToolReadOnly,
+	}
 }
