+++ mastracode/src/harness/harness.ts
@@ -8,6 +8,7 @@ import type { WorkspaceConfig } from "@mastra/core/workspace"
 import type { z } from "zod"
 
 import type {
+	HeartbeatHandler,
 	HarnessConfig,
 	HarnessEvent,
 	HarnessEventListener,
@@ -141,6 +142,10 @@ export class Harness<TState extends HarnessStateSchema = HarnessStateSchema> {
 	private mcpManager: import("../mcp/index.js").MCPManager | undefined
 	private sessionGrants = new SessionGrants()
 	private streamDebug = !!process.env.MASTRA_STREAM_DEBUG
+	private heartbeatTimers = new Map<
+		string,
+		{ timer: NodeJS.Timeout; shutdown?: () => void | Promise<void> }
+	>()
 	private pendingQuestions = new Map<string, (answer: string) => void>()
 	private pendingPlanApprovals = new Map<
 		string,
@@ -258,6 +263,9 @@ export class Harness<TState extends HarnessStateSchema = HarnessStateSchema> {
 				})
 			}
 		}
+
+		// Start heartbeat handlers
+		this.startHeartbeats()
 	}
 
 	/**
@@ -2883,6 +2891,96 @@ export class Harness<TState extends HarnessStateSchema = HarnessStateSchema> {
 	isWorkspaceReady(): boolean {
 		if (this.workspaceFn) return true
 		return this.workspaceInitialized && this.workspace !== undefined
+	}
+
+	// ===========================================================================
+	// Heartbeat Handlers
+	// ===========================================================================
+
+	/**
+	 * Start all configured heartbeat handlers.
+	 * Called automatically during `init()`.
+	 */
+	private startHeartbeats(): void {
+		const handlers = this.config.heartbeatHandlers
+		if (!handlers?.length) return
+
+		for (const hb of handlers) {
+			if (this.heartbeatTimers.has(hb.id)) continue
+
+			const run = async () => {
+				try {
+					await hb.handler()
+				} catch (error) {
+					console.error(`[Heartbeat:${hb.id}] failed:`, error)
+				}
+			}
+
+			if (hb.immediate !== false) {
+				run()
+			}
+
+			const timer = setInterval(run, hb.intervalMs)
+			timer.unref()
+			this.heartbeatTimers.set(hb.id, { timer, shutdown: hb.shutdown })
+		}
+	}
+
+	/**
+	 * Register a heartbeat handler dynamically (after init).
+	 * If a handler with the same id already exists, it is replaced.
+	 */
+	registerHeartbeat(handler: HeartbeatHandler): void {
+		this.removeHeartbeat(handler.id)
+
+		const run = async () => {
+			try {
+				await handler.handler()
+			} catch (error) {
+				console.error(`[Heartbeat:${handler.id}] failed:`, error)
+			}
+		}
+
+		if (handler.immediate !== false) {
+			run()
+		}
+
+		const timer = setInterval(run, handler.intervalMs)
+		timer.unref()
+		this.heartbeatTimers.set(handler.id, { timer, shutdown: handler.shutdown })
+	}
+
+	/**
+	 * Remove a heartbeat handler by id. Calls its shutdown hook if present.
+	 */
+	async removeHeartbeat(id: string): Promise<void> {
+		const entry = this.heartbeatTimers.get(id)
+		if (entry) {
+			clearInterval(entry.timer)
+			this.heartbeatTimers.delete(id)
+			try {
+				await entry.shutdown?.()
+			} catch (error) {
+				console.error(`[Heartbeat:${id}] shutdown failed:`, error)
+			}
+		}
+	}
+
+	/**
+	 * Stop all heartbeat handlers and run their shutdown hooks. Call during shutdown.
+	 */
+	async stopHeartbeats(): Promise<void> {
+		const entries = [...this.heartbeatTimers.entries()]
+		this.heartbeatTimers.clear()
+
+		for (const [id, entry] of entries) {
+			clearInterval(entry.timer)
+			try {
+				await entry.shutdown?.()
+			} catch (error) {
+				console.error(`[Heartbeat:${id}] shutdown failed:`, error)
+			}
+		}
 	}
 
 	/**
