+++ mastracode/src/main.ts
@@ -5,18 +5,16 @@
 import * as fs from "node:fs"
 import * as path from "node:path"
 
-import { createAnthropic } from "@ai-sdk/anthropic"
 import { Mastra } from "@mastra/core"
 import { Agent } from "@mastra/core/agent"
-import { ModelRouterLanguageModel } from "@mastra/core/llm"
 import { noopLogger } from "@mastra/core/logger"
-import type { RequestContext } from "@mastra/core/request-context"
 import { LibSQLStore } from "@mastra/libsql"
-import { Memory } from "@mastra/memory"
-
+import { getDynamicMemory, omState } from "./agents/memory.js"
+import { getDynamicModel } from "./agents/model.js"
+import { createDynamicTools } from "./agents/tools.js"
 import { getDynamicWorkspace } from "./agents/workspace.js"
 import { AuthStorage } from "./auth/storage.js"
-import { DEFAULT_OM_MODEL_ID } from "./constants.js"
+import { DEFAULT_OBS_THRESHOLD, DEFAULT_REF_THRESHOLD } from "./constants.js"
 import { Harness } from "./harness/harness.js"
 import type { HarnessRuntimeContext } from "./harness/types.js"
 import { HookManager } from "./hooks/index.js"
@@ -24,55 +22,24 @@ import { MCPManager } from "./mcp/index.js"
 import type { PromptContext } from "./prompts/index.js"
 import { buildFullPrompt } from "./prompts/index.js"
 import {
-	opencodeClaudeMaxProvider,
 	setAuthStorage,
 } from "./providers/claude-max.js"
 import {
-	openaiCodexProvider,
 	setAuthStorage as setOpenAIAuthStorage,
 } from "./providers/openai-codex.js"
 import { stateSchema } from "./schema.js"
-import {
-	createViewTool,
-	createExecuteCommandTool,
-	stringReplaceLspTool,
-	astSmartEditTool,
-	createWebSearchTool,
-	createWebExtractTool,
-	hasTavilyKey,
-	createGrepTool,
-	createGlobTool,
-	createWriteFileTool,
-	createSubagentTool,
-	todoWriteTool,
-	todoCheckTool,
-	askUserTool,
-	submitPlanTool,
-	requestSandboxAccessTool,
-} from "./tools/index.js"
 import { MastraTUI } from "./tui/index.js"
 import { mastra } from "./tui/theme.js"
-import { startGatewaySync } from "./utils/gateway-sync.js"
+import { syncGateways } from "./utils/gateway-sync.js"
 import {
 	detectProject,
 	getStorageConfig,
 	getUserId,
-	getOmScope,
 	getResourceIdOverride,
 	getAppDataDir,
 } from "./utils/project.js"
 import { releaseAllThreadLocks } from "./utils/thread-lock.js"
 
-
-// =============================================================================
-// Start Gateway Sync (keeps model registry up to date)
-// =============================================================================
-
-startGatewaySync(5 * 60 * 1000) // Sync every 5 minutes
-
-// =============================================================================
-// Helpers
-// =============================================================================
 
 // =============================================================================
 // Create Auth Storage (shared with Claude Max provider and Harness)
@@ -123,199 +90,13 @@ const storage = new LibSQLStore({
 	url: storageConfig.url,
 	...(storageConfig.authToken ? { authToken: storageConfig.authToken } : {}),
 })
-// =============================================================================
-// Create Memory with Observational Memory support
-// =============================================================================
 
-// Default OM thresholds — per-thread overrides are loaded from thread metadata
-const DEFAULT_OBS_THRESHOLD = 40_000
-const DEFAULT_REF_THRESHOLD = 50_000
-
-// Mutable OM state — updated by harness event listeners, read by OM config
-// functions. We use this instead of requestContext because Mastra's OM system
-// does NOT propagate requestContext to observer/reflector agent.generate() calls.
-const omState = {
-	observerModelId: DEFAULT_OM_MODEL_ID,
-	reflectorModelId: DEFAULT_OM_MODEL_ID,
-	obsThreshold: DEFAULT_OBS_THRESHOLD,
-	refThreshold: DEFAULT_REF_THRESHOLD,
-}
-
-/**
- * Dynamic model function for Observer agent.
- * Reads from module-level omState (kept in sync by harness events).
- */
-function getObserverModel() {
-	return resolveModel(omState.observerModelId)
-}
-
-/**
- * Dynamic model function for Reflector agent.
- * Reads from module-level omState (kept in sync by harness events).
- */
-function getReflectorModel() {
-	return resolveModel(omState.reflectorModelId)
-}
-// Cache for Memory instances by threshold config
-let cachedMemory: Memory | null = null
-let cachedMemoryKey: string | null = null
-
-// Resolved OM scope (read once at startup, can be changed via config)
-const omScope = getOmScope(project.rootPath)
-
-/**
- * Dynamic memory factory function.
- * Creates Memory with current threshold values from harness state.
- * Caches instance and reuses if config unchanged.
- */
-function getDynamicMemory({
-	requestContext,
-}: {
-	requestContext: RequestContext
-}) {
-	const ctx = requestContext.get("harness") as
-		| HarnessRuntimeContext<typeof stateSchema>
-		| undefined
-	const state = ctx?.getState?.()
-
-	const obsThreshold = state?.observationThreshold ?? omState.obsThreshold
-	const refThreshold = state?.reflectionThreshold ?? omState.refThreshold
-
-	const cacheKey = `${obsThreshold}:${refThreshold}:${omScope}`
-	if (cachedMemory && cachedMemoryKey === cacheKey) {
-		return cachedMemory
-	}
-
-	cachedMemory = new Memory({
-		storage,
-		options: {
-			observationalMemory: {
-				enabled: true,
-				scope: omScope,
-				observation: {
-					bufferTokens: 1 / 10,
-					bufferActivation: 4 / 5,
-					model: getObserverModel,
-					messageTokens: obsThreshold,
-					blockAfter: 1,
-					modelSettings: {
-						maxOutputTokens: 60000,
-					},
-				},
-				reflection: {
-					bufferActivation: 1 / 2,
-					blockAfter: 1.1,
-					model: getReflectorModel,
-					observationTokens: refThreshold,
-					modelSettings: {
-						maxOutputTokens: 60000,
-					},
-				},
-			},
-		},
-	})
-	cachedMemoryKey = cacheKey
-
-	return cachedMemory
-}
+const memory = getDynamicMemory(storage)
 
 // =============================================================================
-// Create Agent
+// Create MCP Manager
 // =============================================================================
-// Create tools with project root
-const viewTool = createViewTool(project.rootPath)
-const executeCommandTool = createExecuteCommandTool(project.rootPath)
-const grepTool = createGrepTool(project.rootPath)
-const globTool = createGlobTool(project.rootPath)
-const writeFileTool = createWriteFileTool(project.rootPath)
-const webSearchTool = createWebSearchTool()
-const webExtractTool = createWebExtractTool()
-/**
- * Resolve a model ID to the correct provider instance.
- * Shared by the main agent, observer, and reflector.
- *
- * - For anthropic/* models: Uses Claude Max OAuth provider (opencode auth)
- * - For openai/* models with OAuth: Uses OpenAI Codex OAuth provider
- * - For moonshotai/* models: Uses Moonshot AI Anthropic-compatible endpoint
- * - For all other providers: Uses Mastra's model router (models.dev gateway)
- */
-function resolveModel(modelId: string) {
-	const isAnthropicModel = modelId.startsWith("anthropic/")
-	const isOpenAIModel = modelId.startsWith("openai/")
-	const isMoonshotModel = modelId.startsWith("moonshotai/")
-
-	if (isMoonshotModel) {
-		if (!process.env.MOONSHOT_AI_API_KEY) {
-			throw new Error(`Need MOONSHOT_AI_API_KEY`)
-		}
-		return createAnthropic({
-			apiKey: process.env.MOONSHOT_AI_API_KEY!,
-			baseURL: "https://api.moonshot.ai/anthropic/v1",
-			name: "moonshotai.anthropicv1",
-		})(modelId.substring("moonshotai/".length))
-	} else if (isAnthropicModel) {
-		return opencodeClaudeMaxProvider(modelId.substring(`anthropic/`.length))
-	} else if (isOpenAIModel && authStorage.isLoggedIn("openai-codex")) {
-		return openaiCodexProvider(modelId.substring(`openai/`.length))
-	} else {
-		return new ModelRouterLanguageModel(modelId)
-	}
-}
-
-/**
- * Dynamic model function that reads the current model from harness state.
- * This allows runtime model switching via the /models picker.
- */
-function getDynamicModel({
-	requestContext,
-}: {
-	requestContext: RequestContext
-}) {
-	const harnessContext = requestContext.get("harness") as
-		| HarnessRuntimeContext<typeof stateSchema>
-		| undefined
-
-	const modelId = harnessContext?.state?.currentModelId
-	if (!modelId) {
-		throw new Error("No model selected. Use /models to select a model first.")
-	}
-
-	return resolveModel(modelId)
-}
-
-// =============================================================================
-// Create Subagent Tool (subagent delegation)
-// =============================================================================
-
-// The subagent tool needs tools and resolveModel to spawn subagents.
-// We pass all tools that subagents might need based on their type.
-const subagentTool = createSubagentTool({
-	tools: {
-		// Read-only tools (for explore, plan)
-		view: viewTool,
-		search_content: grepTool,
-		find_files: globTool,
-		// Write tools (for execute)
-		string_replace_lsp: stringReplaceLspTool,
-		write_file: writeFileTool,
-		execute_command: executeCommandTool,
-		// Task tracking (for execute)
-		todo_write: todoWriteTool,
-		todo_check: todoCheckTool,
-	},
-	resolveModel,
-})
-
-// Read-only subagent tool for plan mode — no execute type allowed
-const subagentToolReadOnly = createSubagentTool({
-	tools: {
-		view: viewTool,
-		search_content: grepTool,
-		find_files: globTool,
-	},
-	resolveModel,
-	allowedAgentTypes: ["explore", "plan"],
-})
+const mcpManager = new MCPManager(project.rootPath)
 
 // Create agent with dynamic model, dynamic prompt, and full toolset
 const codeAgent = new Agent({
@@ -333,12 +114,12 @@ const codeAgent = new Agent({
 			projectName: state?.projectName ?? project.name,
 			gitBranch: state?.gitBranch ?? project.gitBranch,
 			platform: process.platform,
-			date: new Date().toISOString().split("T")[0],
+			date: new Date().toISOString().split("T")[0]!,
 			mode: modeId,
 			activePlan: state?.activePlan ?? null,
 			// Add missing fields for PromptContext
 			modeId: modeId,
-			currentDate: new Date().toISOString().split("T")[0],
+			currentDate: new Date().toISOString().split("T")[0]!,
 			workingDir: state?.projectPath ?? project.rootPath,
 			state: state,
 		}
@@ -346,61 +127,9 @@ const codeAgent = new Agent({
 		return buildFullPrompt(promptCtx)
 	},
 	model: getDynamicModel,
-	memory: getDynamicMemory,
+	memory,
 	workspace: getDynamicWorkspace,
-	tools: ({ requestContext }) => {
-		const harnessContext = requestContext.get("harness") as
-			| HarnessRuntimeContext<typeof stateSchema>
-			| undefined
-		const modeId = harnessContext?.modeId ?? "build"
-
-		// Build tool set based on mode
-		// NOTE: Tool names "grep" and "glob" are reserved by Anthropic's OAuth
-		// validation (they match Claude Code's internal tools). We use
-		// "search_content" and "find_files" to avoid the collision.
-		const tools: Record<string, any> = {
-			// Read-only tools — always available
-			view: viewTool,
-			search_content: grepTool,
-			find_files: globTool,
-			execute_command: executeCommandTool,
-			// Subagent delegation — read-only in plan mode
-			subagent: modeId === "plan" ? subagentToolReadOnly : subagentTool,
-			// Todo tracking — always available (planning tool, not a write tool)
-			todo_write: todoWriteTool,
-			todo_check: todoCheckTool,
-			// User interaction — always available
-			ask_user: askUserTool,
-			request_sandbox_access: requestSandboxAccessTool,
-		}
-
-		// Write tools — NOT available in plan mode
-		if (modeId !== "plan") {
-			tools.string_replace_lsp = stringReplaceLspTool
-			tools.ast_smart_edit = astSmartEditTool
-			tools.write_file = writeFileTool
-		}
-
-		// Plan submission — only available in plan mode
-		if (modeId === "plan") {
-			tools.submit_plan = submitPlanTool
-		}
-		// Web tools — prefer Tavily when available (avoids Anthropic native
-		// web_search provider tool which can cause stream freezes). Fall back
-		// to Anthropic's native web search via getToolsets() for Anthropic models.
-		// Note: hasTavilyKey() is checked at request time, not module load time,
-		// so the key can be set after startup and still be picked up.
-		if (hasTavilyKey()) {
-			tools.web_search = webSearchTool
-			tools.web_extract = webExtractTool
-		}
-
-		// MCP server tools — injected from connected servers
-		const mcpTools = mcpManager.getTools()
-		Object.assign(tools, mcpTools)
-
-		return tools
-	},
+	tools: createDynamicTools(mcpManager),
 })
 
 // Register the agent with a Mastra instance so that workflow snapshot storage
@@ -418,28 +147,6 @@ mastraInstance.getLogger = () => noopLogger as any
 codeAgent.__setLogger(noopLogger)
 
 // =============================================================================
-// Anthropic Provider Tools (web search & fetch - zero implementation needed)
-// Only used as fallback when Tavily API key is not configured.
-// =============================================================================
-const anthropic = createAnthropic({})
-
-function getToolsets(
-	modelId: string,
-): Record<string, Record<string, unknown>> | undefined {
-	// If Tavily is available, skip Anthropic's native web search
-	if (hasTavilyKey()) return undefined
-
-	const isAnthropicModel = modelId.startsWith("anthropic/")
-	if (!isAnthropicModel) return undefined
-
-	return {
-		anthropic: {
-			web_search: anthropic.tools.webSearch_20250305(),
-		},
-	}
-}
-
-// =============================================================================
 // Create Hook Manager
 // =============================================================================
 const hookManager = new HookManager(project.rootPath, "session-init")
@@ -452,11 +159,6 @@ if (hookManager.hasHooks()) {
 	)
 	console.info(`Hooks: ${hookCount} hook(s) configured`)
 }
-
-// =============================================================================
-// Create MCP Manager
-// =============================================================================
-const mcpManager = new MCPManager(project.rootPath)
 
 // =============================================================================
 // Create Harness
@@ -474,7 +176,6 @@ const harness = new Harness({
 		projectName: project.name,
 		gitBranch: project.gitBranch,
 	},
-	getToolsets,
 	workspace: getDynamicWorkspace,
 	hookManager,
 	mcpManager,
@@ -503,6 +204,13 @@ const harness = new Harness({
 		},
 	],
 	authStorage, // Share auth storage with Claude Max provider
+	heartbeatHandlers: [
+		{
+			id: "gateway-sync",
+			intervalMs: 5 * 60 * 1000,
+			handler: () => syncGateways(),
+		},
+	],
 })
 
 // Keep omModelState in sync with harness state changes.
@@ -589,22 +297,25 @@ const tui = new MastraTUI({
 			process.exit(1)
 		})
 	})()
-// Clean up MCP connections on exit
+// Graceful async cleanup (MCP connections, heartbeat shutdown hooks)
 process.on("beforeExit", async () => {
-	await mcpManager.disconnect()
+	await Promise.all([
+		mcpManager.disconnect(),
+		harness.stopHeartbeats(),
+	])
 })
 
-// Release thread locks on exit (handles SIGINT, SIGTERM, and normal exit)
-const cleanupThreadLocks = () => {
+// Synchronous cleanup on exit / signals (thread locks)
+const cleanup = () => {
 	harness.releaseCurrentThreadLock()
 	releaseAllThreadLocks()
 }
-process.on("exit", cleanupThreadLocks)
+process.on("exit", cleanup)
 process.on("SIGINT", () => {
-	cleanupThreadLocks()
+	cleanup()
 	process.exit(0)
 })
 process.on("SIGTERM", () => {
-	cleanupThreadLocks()
+	cleanup()
 	process.exit(0)
 })
