+++ mastracode/src/utils/gateway-sync.ts
@@ -4,8 +4,9 @@
  */
 
 import fs from "node:fs"
-import path from "node:path"
 import os from "node:os"
+import path from "node:path"
+
 import { ModelsDevGateway, NetlifyGateway } from "@mastra/core/llm"
 import type { ProviderConfig } from "@mastra/core/llm"
